ci: validate native Flutter builds and fix dashboard types (#39)
* fix: use native Gemini provider for BYOK runtime

* fix: align Gemini provider package with AI SDK runtime

* fix: align Radix aspect-ratio with published version

* fix: align remaining radix dependency with published version

* fix: correct radix collapsible version without unrelated dependency changes

* fix: restore package manifest while keeping radix fix

* fix: keep manifest changes limited to published radix version

* fix: align radix dropdown menu dependency

* fix: use published radix dropdown menu version

* fix: normalize dashboard clamp result to double

* fix: normalize dashboard progress values to double

* ci: run native Flutter validation on pull requests

+++ pace_flutter/lib/features/dashboard/dashboard_model.dart
@@ -63,7 +63,7 @@ class DashboardSnapshot {
   final Map<String, int> routineDoneByDay;
   final Map<String, double> weightByDay;
 
-  double _ratio(double value, double target) => target <= 0 ? 0 : (value / target).clamp(0, 1);
+  double _ratio(double value, double target) => target <= 0 ? 0 : (value / target).clamp(0, 1).toDouble();
 
   int scoreFor(String day) {
     final routineTotalSafe = routineTotal == 0 ? 1 : routineTotal;
